style:rename btn
